+++ zcash_client_sqlite/src/wallet/init.rs
@@ -134,11 +134,10 @@ fn sqlite_client_error_to_wallet_migration_error(e: SqliteClientError) -> Wallet
         SqliteClientError::InvalidNote => {
             WalletMigrationError::CorruptedData("invalid note".into())
         }
-        SqliteClientError::Bech32DecodeError(e) => {
-            WalletMigrationError::CorruptedData(e.to_string())
-        }
+        SqliteClientError::DecodingError(e) => WalletMigrationError::CorruptedData(e.to_string()),
         #[cfg(feature = "transparent-inputs")]
         SqliteClientError::HdwalletError(e) => WalletMigrationError::CorruptedData(e.to_string()),
+        #[cfg(feature = "transparent-inputs")]
         SqliteClientError::TransparentAddress(e) => {
             WalletMigrationError::CorruptedData(e.to_string())
         }